+++ examples/basic_authentication/analysis_options.yaml
@@ -1,4 +1,4 @@
-include: package:very_good_analysis/analysis_options.4.0.0.yaml
+include: package:very_good_analysis/analysis_options.5.0.0.yaml
 analyzer:
   exclude:
     - build/**
