+++ examples/basic_authentication/pubspec.yaml
@@ -16,4 +16,4 @@ dependencies:
 dev_dependencies:
   mocktail: ^0.3.0
   test: ^1.19.2
-  very_good_analysis: ^4.0.0
+  very_good_analysis: ^5.0.0
